move App and affected components into src/
