+++ protocol-designer/src/ProtocolEditor.tsx
@@ -11,26 +11,26 @@ import {
   PrimaryButton,
   SPACING,
 } from '@opentrons/components'
-import { getEnableRedesign } from '../feature-flags/selectors'
-import { setFeatureFlags } from '../feature-flags/actions'
-import { ComputingSpinner } from '../components/ComputingSpinner'
-import { ConnectedNav } from '../containers/ConnectedNav'
-import { Sidebar } from '../containers/ConnectedSidebar'
-import { ConnectedTitleBar } from '../containers/ConnectedTitleBar'
-import { MainPanel } from '../containers/ConnectedMainPanel'
-import { PortalRoot as MainPageModalPortalRoot } from '../components/portals/MainPageModalPortal'
-import { MAIN_CONTENT_FORCED_SCROLL_CLASSNAME } from '../ui/steps/utils'
-import { PrereleaseModeIndicator } from './PrereleaseModeIndicator'
-import { PortalRoot as TopPortalRoot } from './portals/TopPortal'
-import { FileUploadMessageModal } from './modals/FileUploadMessageModal/FileUploadMessageModal'
-import { LabwareUploadMessageModal } from './modals/LabwareUploadMessageModal/LabwareUploadMessageModal'
-import { GateModal } from './modals/GateModal'
-import { CreateFileWizard } from './modals/CreateFileWizard'
-import { AnnouncementModal } from './modals/AnnouncementModal'
+import { getEnableRedesign } from './feature-flags/selectors'
+import { setFeatureFlags } from './feature-flags/actions'
+import { ComputingSpinner } from './components/ComputingSpinner'
+import { ConnectedNav } from './containers/ConnectedNav'
+import { Sidebar } from './containers/ConnectedSidebar'
+import { ConnectedTitleBar } from './containers/ConnectedTitleBar'
+import { MainPanel } from './containers/ConnectedMainPanel'
+import { PortalRoot as MainPageModalPortalRoot } from './components/portals/MainPageModalPortal'
+import { MAIN_CONTENT_FORCED_SCROLL_CLASSNAME } from './ui/steps/utils'
+import { PrereleaseModeIndicator } from './components/PrereleaseModeIndicator'
+import { PortalRoot as TopPortalRoot } from './components/portals/TopPortal'
+import { FileUploadMessageModal } from './components/modals/FileUploadMessageModal/FileUploadMessageModal'
+import { LabwareUploadMessageModal } from './components/modals/LabwareUploadMessageModal/LabwareUploadMessageModal'
+import { GateModal } from './components/modals/GateModal'
+import { CreateFileWizard } from './components/modals/CreateFileWizard'
+import { AnnouncementModal } from './components/modals/AnnouncementModal'
 import { ProtocolRoutes } from './ProtocolRoutes'
 
-import styles from './ProtocolEditor.module.css'
-import '../css/reset.module.css'
+import styles from './components/ProtocolEditor.module.css'
+import './css/reset.module.css'
 
 const showGateModal =
   process.env.NODE_ENV === 'production' || process.env.OT_PD_SHOW_GATE
