+++ protocol-designer/src/ProtocolRoutes.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react'
 import { Route, Navigate, Routes, useLocation } from 'react-router-dom'
 import { Box } from '@opentrons/components'
-import { Landing } from '../pages/Landing'
-import { ProtocolOverview } from '../pages/ProtocolOverview'
-import { Liquids } from '../pages/Liquids'
-import { StartingDeckState } from '../pages/StartingDeckState'
-import { ProtocolSteps } from '../pages/ProtocolSteps'
-import { CreateNewProtocol } from '../pages/CreateNewProtocol'
+import { Landing } from './pages/Landing'
+import { ProtocolOverview } from './pages/ProtocolOverview'
+import { Liquids } from './pages/Liquids'
+import { StartingDeckState } from './pages/StartingDeckState'
+import { ProtocolSteps } from './pages/ProtocolSteps'
+import { CreateNewProtocol } from './pages/CreateNewProtocol'
 import { Navbar } from './Navbar'
 
 import type { RouteProps } from './types'
