+++ protocol-designer/src/index.tsx
@@ -4,10 +4,10 @@ import { Provider } from 'react-redux'
 import { I18nextProvider } from 'react-i18next'
 
 import { configureStore } from './configureStore'
-import { App } from './components/App'
 import { initialize } from './initialize'
 import { initializeMixpanel } from './analytics/mixpanel'
 import { i18n } from './localization'
+import { App } from './App'
 
 // initialize Redux
 const store = configureStore()
